+++ lib/liveValidation/liveValidator.ts
@@ -843,7 +843,8 @@ export class LiveValidator {
       elapsedTime,
       info.validationRequest
     );
-    console.log(`TIme ${elapsedTime} to validate ${operationId}`);
+    delete info.validationRequest;
+    delete info.operationMatch;
     return {
       isSuccessful: runtimeException ? undefined : errors.length === 0,
       operationInfo: info,