+++ ChangeLog.md
@@ -9,6 +9,8 @@
   - Improve LRO poller and use retry-after header in delay
   - Improve logging level to reuse the `-l` option
   - Clean some unused dependencies
+  - Support Roundtrip validation in validation
+  - Integrate Roundtrip validation into API Scenario
 
 ## 08/04/2022 3.0.5